debugging

+++ src/main.cpp
@@ -60,8 +60,10 @@ int main(int argc, char *argv[])
         model.create_full_model(data);
         model.execute_solver_for_full_model(data);
         model.get_value_of_variables(data);
-        std::cout << std::endl << "-> Total time = " << model.current_sol.computational_time.count() << std::endl;
+        std::cout << std::endl << "-> Total time = " << double(model.current_sol.computational_time.count()) << std::endl;
         model.current_sol.display_solution(data);
+        // model.current_sol.display_variables_values(data);
+        // model.current_sol.create_graph(data, method, 1);
     }
     else if (method == "enum")
     {
